First raw upgrade: PBR scene, iPhone support

+++ ExploringSceneKit/Model/SceneFactory.h
@@ -13,6 +13,6 @@
 
 @interface SceneFactory : NSObject
 
-- (SCNScene<Scene> *)makeScene:(int)sceneId;
+- (SCNScene<Scene> *)makeScene:(NSInteger)sceneId;
 
 @end
